requirements.txt
streamlit
transformers
torch
Pillow
numpy
sentencepiece